Move opcodes into types and decoupled parts of implementation from other modules
